Save production project

+++ pepe-ganga/consuming-webservice-store/src/main/java/uy/com/pepeganga/consumingwsstore/repositories/ICategoryRepository.java
@@ -7,6 +7,6 @@ import uy.com.pepeganga.business.common.entities.*;
 
 
 @Repository
-public interface ICategoryRepository extends JpaRepository<Category, Integer>{
+public interface ICategoryRepository extends JpaRepository<Category, Short>{
 
 }
